DrainRecord no longer has an overlap pixel with other local maps
This guarantees that pixels on the east and south correctly tag as
'offmap' without double counting pixels. Ultimately, it fixes a ton of
bugs in hydrology and throws the entire "overlap" pixel concept into
doubt... although the overlap pixel concept was built on the assumption
that we would be interpolating between pixels, and that a local map
needed the ability to run interpolation on its interior completely
independently from other local maps. This may still be valid for the
DataImage classes, but its' proven bad for creating consistent
hydrology.