Basic Rendering!
Adding a mediocre triangle rasterizer, and associated "photographer"
class, so that we can actually do a 3D rendering of the terrain we're
generating. Definitely seeing the limitations of doing this in java,
where we don't have good access to flat data buffers... maybe I'll
switch to c++ one of these days.